Simplified connection manager API (#225)
Replaced implementation of connection manager.

The original implementation of the connection manager was made with the assumption that multiple requests could be made over the same connection. Because of that it was based off of context managers. Since SPAMD will close the connection once the response is sent there's no need to maintain a context for it.

This removes the context managers and simplifies the API for the connection.